refactor: Merge parsing map part and playing the game part#26
- 기존 따로 작업했던 부분을 합쳤다.
- 목적 파일을 다른 디렉토리에 만들던 작업을 소스 코드와 함께 만들도록
  변경했다.
- so long 때 사용했던 parser 부분 삭제하고 내가 작업하던 parse_map 으로
  교체했다.
