+++ Makefile
@@ -3,10 +3,10 @@
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: kipark <kipark@student.42seoul.kr>         +#+  +:+       +#+         #
+#    By: jayoon <jayoon@student.42.fr>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/04/27 15:13:41 by kipark            #+#    #+#              #
-#    Updated: 2022/11/16 20:28:53 by kipark           ###   ########seoul.kr   #
+#    Updated: 2022/11/17 16:58:10 by jayoon           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -21,9 +21,9 @@ LIBFT_FLAGS 				=	-L$(DIR_LIBFT) -lft
 INCS_FLAGS					=	-I $(DIR_LIBFT) -I $(DIR_INCS)
 LIBFT						=	$(DIR_LIBFT)/libft.a
 
-MLXLIB 						= $(DIR_MLX)/libmlx.a
-DIR_MLX 					= ./minilibx_opengl_20191021
-MLX_FLAGS 					= -L$(DIR_MLX) -lmlx -framework OpenGL -framework AppKit
+MLXLIB 						= 	$(DIR_MLX)/libmlx.a
+DIR_MLX 					= 	./minilibx_opengl_20191021
+MLX_FLAGS 					= 	-L$(DIR_MLX) -lmlx -framework OpenGL -framework AppKit
 
 DIR_SRCS					:=	./srcs
 DIR_INCS 					:=	./includes
@@ -34,64 +34,57 @@ DIR_EVENT					:=	$(DIR_SRCS)/event
 DIR_ERROR					:=	$(DIR_SRCS)/error
 DIR_MINI_MAP				:=	$(DIR_SRCS)/mini_map
 DIR_GAME_VIEW				:=	$(DIR_SRCS)/game_view
-DIR_PARSER					:=	$(DIR_SRCS)/parser
+DIR_PARSE_MAP				:=	$(DIR_SRCS)/parse_map
 DIR_SET_GAME				:=	$(DIR_SRCS)/set_game
 
 SRCS_MAIN					:=	main.c
-OJBS_MAIN					:=	$(addprefix $(DIR_OBJS)/, $(SRCS_MAIN))
 SRCS_MAIN					:=	$(addprefix $(DIR_SRCS)/, $(SRCS_MAIN))
 
-SRCS_PARSER					:=	error_utils.c					\
-								parse_error.c					\
-								parse_utils.c					\
-								parse.c
-OJBS_PARSER					:=	$(addprefix $(DIR_OBJS)/, $(SRCS_PARSER))
-SRCS_PARSER					:=	$(addprefix $(DIR_PARSER)/, $(SRCS_PARSER))
+SRCS_PARSE_MAP				:=	init_map_info.c			\
+								cub3d_atoi.c			\
+								init_element.c			\
+								process_color.c 		\
+								process_texture_path.c	\
+								init_map_content.c		\
+								safe_open.c				\
+								get_cnt_map_height.c	\
+								get_cnt_map_width.c		\
+								pass_empty_line.c		\
+								pass_element.c			\
+								get_map.c				\
+								is_valid_map.c			\
+								is_valid_character.c
+SRCS_PARSE_MAP				:=	$(addprefix $(DIR_PARSE_MAP)/, $(SRCS_PARSE_MAP))
 
 SRCS_EVENT					:=	key.c
-OJBS_EVENT					:=	$(addprefix $(DIR_OBJS)/, $(SRCS_EVENT))
 SRCS_EVENT					:=	$(addprefix $(DIR_EVENT)/, $(SRCS_EVENT))
 
 SRCS_MINI_MAP				:=	draw_mini_map.c			\
 								draw_mini_player.c		\
 								draw_mini_ray.c
-OJBS_MINI_MAP				:=	$(addprefix $(DIR_OBJS)/, $(SRCS_MINI_MAP))
 SRCS_MINI_MAP				:=	$(addprefix $(DIR_MINI_MAP)/, $(SRCS_MINI_MAP))
 
 SRCS_ERROR					:=	error.c
-OJBS_ERROR					:=	$(addprefix $(DIR_OBJS)/, $(SRCS_ERROR))
 SRCS_ERROR					:=	$(addprefix $(DIR_ERROR)/, $(SRCS_ERROR))
 
 SRCS_GAME_VIEW				:=	game_view.c 			\
 								game_dda.c				\
 								game_draw.c
-OJBS_GAME_VIEW				:=	$(addprefix $(DIR_OBJS)/, $(SRCS_GAME_VIEW))
 SRCS_GAME_VIEW				:=	$(addprefix $(DIR_GAME_VIEW)/, $(SRCS_GAME_VIEW))
 
 SRCS_SET_GAME				:=	set_mlx.c				\
 								set_texture.c
-OJBS_SET_GAME				:=	$(addprefix $(DIR_OBJS)/, $(SRCS_SET_GAME))
 SRCS_SET_GAME				:=	$(addprefix $(DIR_SET_GAME)/, $(SRCS_SET_GAME))
 
 SRCS						:=	$(SRCS_MAIN)			\
 								$(SRCS_EVENT)			\
 								$(SRCS_ERROR)			\
-								$(SRCS_PARSER)			\
+								$(SRCS_PARSE_MAP)		\
 								$(SRCS_MINI_MAP)		\
 								$(SRCS_GAME_VIEW)		\
 								$(SRCS_SET_GAME)
 								
-
 OBJS						:=	$(SRCS:.c=.o)
-OBJS_SRCS					:= 	$(OJBS_MAIN)			\
-								$(OJBS_EVENT)			\
-								$(OJBS_ERROR)			\
-								$(OJBS_PARSER)			\
-								$(OJBS_MINI_MAP)		\
-								$(OJBS_GAME_VIEW)		\
-								$(OJBS_SET_GAME)
-
-OBJS_OBJS					:=	$(OBJS_SRCS:.c=.o)
 
 ifdef DEBUG_FLAG
 	CFLAGS += $(DEBUG)
@@ -104,18 +97,16 @@ endif
 all: $(NAME)
 
 $(NAME)		: $(LIBFT) $(MLXLIB) $(OBJS)
-	$(CC) $(CFLAGS) $(LIBFT_FLAGS) $(MLX_FLAGS) $(OBJS_OBJS) -o $@
-	./cub3d ./map/test_map.ber
+	$(CC) $(CFLAGS) $(LIBFT_FLAGS) $(MLX_FLAGS) $(OBJS) -o $@
 
 %.o			: %.c
 	$(CC) $(CFLAGS) $(INCS_FLAGS)  -c $< -o $@
-	mv $@ ./objs
-
-$(MLXLIB)	:
-	make -C $(DIR_MLX) all
 
 $(LIBFT)	:
 	make -C $(DIR_LIBFT) all
+
+$(MLXLIB)	:
+	make -C $(DIR_MLX) all
 
 debug: 
 	$(MAKE) DEBUG_FLAG=1
@@ -132,7 +123,6 @@ fclean:
 	make clean
 	make fclean -C $(DIR_LIBFT)
 	make fclean -C $(DIR_MLX)
-	rm -f $(DIR_OBJS)/*
 	rm -rf $(NAME)
 
 re:
