+++ srcs/game_view/game_dda.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   game_dda.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kipark <kipark@student.42seoul.kr>         +#+  +:+       +#+        */
+/*   By: jayoon <jayoon@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/16 16:59:38 by kipark            #+#    #+#             */
-/*   Updated: 2022/11/16 22:03:23 by kipark           ###   ########seoul.kr  */
+/*   Updated: 2022/11/17 16:47:07 by jayoon           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
