--- srcs/img/e.xpm
@@ -1,160 +1,0 @@
-/* XPM */
-static char *result[] = {
-/* columns rows colors chars-per-pixel */
-"64 64 90 1 ",
-"  c #4D3E40",
-". c #4F4042",
-"X c #544445",
-"o c #5C4642",
-"O c #594947",
-"+ c #564648",
-"@ c #58474A",
-"# c #564A4A",
-"$ c #5A4A4C",
-"% c #5A514F",
-"& c #5E4D50",
-"* c #5C5451",
-"= c #5F5955",
-"- c #664C43",
-"; c #6C4E45",
-": c #634E4B",
-"> c #6B544E",
-", c #67584E",
-"< c #735344",
-"1 c #73584E",
-"2 c #614F52",
-"3 c #645255",
-"4 c #615B56",
-"5 c #675458",
-"6 c #695659",
-"7 c #635F59",
-"8 c #6C595C",
-"9 c #705852",
-"0 c #7D5E51",
-"q c #705C5F",
-"w c #65615B",
-"e c #68665F",
-"r c #6F5C60",
-"t c #735F63",
-"y c #6C6B63",
-"u c #6F7167",
-"i c #72746A",
-"p c #75796D",
-"a c #777D70",
-"s c #797F72",
-"d c #896656",
-"f c #84745E",
-"g c #A16A4A",
-"h c #867560",
-"j c #8D7D63",
-"k c #A97A61",
-"l c #7B8275",
-"z c #7E8778",
-"x c #7F8979",
-"c c #948467",
-"v c #968668",
-"b c #838D7D",
-"n c #BF8664",
-"m c #BD8868",
-"M c #A98370",
-"N c #A59571",
-"B c #C68E6B",
-"V c #869382",
-"C c #889583",
-"Z c #8B9A87",
-"A c #8F9F8B",
-"S c #92A58F",
-"D c #93A691",
-"F c #97AC95",
-"G c #98AE96",
-"H c #9BB29A",
-"J c #9FB89E",
-"K c #A3BEA3",
-"L c #DBA98A",
-"P c #C3B381",
-"I c #CAB985",
-"U c #E3AF8E",
-"Y c #ECB28C",
-"T c #EDB693",
-"R c #A6C3A6",
-"E c #A8C5A9",
-"W c #ABCAAD",
-"Q c #AFCFB1",
-"! c #AFD1B1",
-"~ c #B2D4B4",
-"^ c #B5D8B6",
-"/ c #B6DAB9",
-"( c #B9DEBC",
-") c #BAE0BD",
-"_ c #D7C78D",
-"` c #DDCC90",
-"' c #FFC49C",
-"] c #F3E29C",
-"[ c #FFEEA3",
-"{ c None",
-/* pixels */
-"{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{",
-"{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{",
-"{{{{{{{{gggggggggggggggggggggggggggggggggggggggggggggggg{{{{{{{{",
-"{{{{{{{{gnYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYng{{{{{{{{",
-"{{{{{{{{gB'ULLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLU'Bg{{{{{{{{",
-"{{{{{{{{gB'9                                        9'Bg{{{{{{{{",
-"{{{{{{{{gB'9X33333333333333333333333333333333366333X9'Bg{{{{{{{{",
-"{{{{{{{{gB'9XttttttttttttttttttttttttttttttttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9Xtttttttt3@ .@&rtttttttttttttttttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9Xtttttt8X ,jj, Xrtttttttttttttttttttttt@9'Bg{{{{{{{{",
-"{{{{{{{{gB'9Xttttt8XOI[[[[IOX8tttttttttttttttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9Xttttt@X`[[[[[[`X$tttttttttttttttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9Xttttr v[[[[[[[[c rttttttttttttttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9+tttt6 I[[[[[[[[I 3ttttttttttttttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9+tttt5 _[[[[[[[[_ 3ttttttttttttttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9Xtttt3 P[[[[[[[[P 6tttttttttttttttttttt+9'Bg{{{{{{{{",
-"{{{{{{{{gB'9XtttttXf[[[[[[[[f.ttttttttttttttttttttt+9'Bg{{{{{{{{",
-"{{{{{{{{gB'9Xttttt$.P[[[[[[P &tttttttttttttttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9XttttttX.N][[]N.@rttttttttttttttttttttt+9'Bg{{{{{{{{",
-"{{{{{{{{gB'9Xttttttt@ .,,. @rttttttttttttttttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9+tttttttt6$@X$8tttqrt8tyrrrrtttttttttttO1'Bg{{{{{{{{",
-"{{{{{{{{gB'9Ottttttttttttt83@.       ...X$8ttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9Xttttttttttq&X  #pAAAAAAAAAAb*.3ttttttt+9'Bg{{{{{{{{",
-"{{{{{{{{gB'1XttttttttttX.yZR))))(((()))))/w.rtttttt+9'Bg{{{{{{{{",
-"{{{{{{{{gB'9+ttttttttt@ A)))))))))))))))))E.$ttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9Xtttttttt$ b)))))))))))()))))))x.6tttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9Xttttttt$ p()))))))))()((()))))/*XrttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9Xtttttt3 y())))))))))^w7777a))))F &ttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'>X668883 4^))))))))))))u XX  G))))yXqttt@9'Bg{{{{{{{{",
-"{{{{{{{{gB'X      .#Q))))^R))))))(H 2tt %()))EX#tttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'X%AAAAASW))))(yy)))))))(4Xtt3 V))))a.qttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'Xw)))))(()))(i  E))))))(A 6ttXXW)))b qtt+9'Bg{{{{{{{{",
-"{{{{{{{{gB'X7)))))))))(a #.l)))))))^X@tt6 y()~*@tttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'.7))))))))(z #t@.~)))))))z 8tt$ *p#.3tttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'X7)))))))^s Xtt6.V)))))))W $ttt3X @3ttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'XXw77777=. @ttttX#())))))(y ttttttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'O.X@X@XXX$3ttttt3 S())))))G 3tttttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'1XtttttttttttttttX7)))))))/XXtttttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9Xttttttttttttttt& J)))))))$.rttttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9@tttttttttttttttr.u))))))(#.&tttttttttt@9'Bg{{{{{{{{",
-"{{{{{{{{gB'9Xtttttttttttttttt$ !)))))Q.X.8tttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9Xtttttttttttttttt.y))))))x.V.#tttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9Xttttttttttttttt& K)))))E.p(8 8ytttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9Xttttttttttttttr.i))))))y.!)K  ttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9$tttttttttttttt&.R)))))S A)))i.38888888X9'Bg{{{{{{{{",
-"{{{{{{{{gB'9@ttttttttttttt8.a)))))~#w()))W.        .OTBg{{{{{{{{",
-"{{{{{{{{gB'9Xttttttttttttt$.Q)))))a Z)))))HAAAAAAAC7 :mg{{{{{{{{",
-"{{{{{{{{gB'9@tttttttttttt8.x)))))J.##!)))))))))))))/7.dg{{{{{{{{",
-"{{{{{{{{gB'9Xtttttttttttt@X!))))(=Xt$y()))))))))))))Z 1g{{{{{{{{",
-"{{{{{{{{gB'9Xttttttttttt8 V)))))V.3t8XC())))))))))))b 0g{{{{{{{{",
-"{{{{{{{{gB'9Xttttttttttt@#^))))W.$ttt3 x~)))))(()))RX kg{{{{{{{{",
-"{{{{{{{{gB'9Xtttttttttt6 Z)))))y.rtttt5X =44444444*  MBg{{{{{{{{",
-"{{{{{{{{gB'9@tttttttttt@%())))D.3ttttttt5&$@$$$$$$$X>'Bg{{{{{{{{",
-"{{{{{{{{gB'9Xttttttttt3 D()))~#@tttttttttttttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9OtttttttttX=())))l 8tttttttttttttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9Xtttttttt3 H))))K.$ttttttttttttttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9XttttttttXw())((=XtttttttttttttttttttttX>'Bg{{{{{{{{",
-"{{{{{{{{gB'9Xttttttt& K))))V 5tttttttttttttttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9Xttttttt.y))))WX@ttttttttttttttttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gB'9Xtttttt& R)))(yX8ttttttttttttttttttttttX9'Bg{{{{{{{{",
-"{{{{{{{{gBY>@868885 eRKKHi &88868888r88888886888888X>YBg{{{{{{{{",
-"{{{{{{{{gggo                                        oggg{{{{{{{{",
-"{{{{{{{{{{{<;;;;;;oooOooo-;;;;;;;;;;;;;;;;;;;;;;;;;;<{{{{{{{{{{{",
-"{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{"
-};
