--- srcs/img/p.xpm
@@ -1,223 +1,0 @@
-/* XPM */
-static char *result[] = {
-/* columns rows colors chars-per-pixel */
-"42 42 175 2 ",
-"   c #483E36",
-".  c #5E3C3A",
-"X  c #643F3C",
-"o  c #6E3F3E",
-"O  c #49412F",
-"+  c #4D4530",
-"@  c #4A4039",
-"#  c #51453E",
-"$  c #5C443F",
-"%  c #534B32",
-"&  c #575132",
-"*  c #5A5333",
-"=  c #5E5935",
-"-  c #5F583C",
-";  c #5B543E",
-":  c #66423F",
-">  c #655A39",
-",  c #726E3D",
-"<  c #76763D",
-"1  c #79743E",
-"2  c #7B793D",
-"3  c #676338",
-"4  c #574A42",
-"5  c #5E5449",
-"6  c #5C5343",
-"7  c #644742",
-"8  c #664E48",
-"9  c #6D4845",
-"0  c #764243",
-"q  c #7D4546",
-"w  c #774C49",
-"e  c #7A4D4B",
-"r  c #724946",
-"t  c #63554C",
-"y  c #635B48",
-"u  c #685B51",
-"i  c #745345",
-"p  c #6D654A",
-"a  c #6A6242",
-"s  c #736D40",
-"d  c #777242",
-"f  c #7D7B43",
-"g  c #6D6255",
-"h  c #736657",
-"j  c #72665C",
-"k  c #796B5E",
-"l  c #76695F",
-"z  c #717C5E",
-"x  c #755D62",
-"c  c #7A6D62",
-"v  c #7B6766",
-"b  c #7F7167",
-"n  c #7F7269",
-"m  c #716D73",
-"M  c #727677",
-"N  c #657469",
-"B  c #84494A",
-"V  c #8B494C",
-"C  c #95484C",
-"Z  c #90464B",
-"A  c #884F50",
-"S  c #8D5453",
-"D  c #975653",
-"F  c #A54A51",
-"G  c #AC4B54",
-"H  c #BB4F58",
-"J  c #A65E58",
-"K  c #AA5C5B",
-"L  c #A75557",
-"P  c #B55656",
-"I  c #8B6449",
-"U  c #917F47",
-"Y  c #857E54",
-"T  c #887245",
-"R  c #A16153",
-"E  c #B8665F",
-"W  c #855F64",
-"Q  c #955D63",
-"!  c #A15F66",
-"~  c #BB5C62",
-"^  c #816267",
-"/  c #83766C",
-"(  c #87796F",
-")  c #837567",
-"_  c #AB6167",
-"`  c #BC616B",
-"'  c #D34E5C",
-"]  c #C4515C",
-"[  c #CC525D",
-"{  c #D3515E",
-"}  c #C35D66",
-"|  c #C65F69",
-" . c #D85461",
-".. c #D85764",
-"X. c #E35E6B",
-"o. c #CD606A",
-"O. c #C7636A",
-"+. c #D3636D",
-"@. c #D9636E",
-"#. c #DD6672",
-"$. c #E6616F",
-"%. c #E46673",
-"&. c #EA6573",
-"*. c #E66975",
-"=. c #EA6A77",
-"-. c #ED6A78",
-";. c #F16C7A",
-":. c #EC7B75",
-">. c #7D895F",
-",. c #7C8556",
-"<. c #748266",
-"1. c #77847E",
-"2. c #83823F",
-"3. c #8A8A41",
-"4. c #878740",
-"5. c #908346",
-"6. c #99994F",
-"7. c #8D8951",
-"8. c #8B8756",
-"9. c #918D56",
-"0. c #8A9150",
-"q. c #949057",
-"w. c #989F53",
-"e. c #A29C4A",
-"r. c #9DA34F",
-"t. c #A4A84F",
-"y. c #AAA94D",
-"u. c #AAB14E",
-"i. c #B8BD4E",
-"p. c #A4A950",
-"a. c #ADB050",
-"s. c #B3B952",
-"d. c #938965",
-"f. c #998B7A",
-"g. c #948772",
-"h. c #BBC14F",
-"j. c #BEC350",
-"k. c #C3CA52",
-"l. c #CDD454",
-"z. c #D4DC56",
-"x. c #D0D755",
-"c. c #D7E057",
-"v. c #738785",
-"b. c #748994",
-"n. c #76949C",
-"m. c #77999D",
-"M. c #789CA9",
-"N. c #78A0A6",
-"B. c #79A5B5",
-"V. c #7CABBD",
-"C. c #818580",
-"Z. c #979183",
-"A. c #A79986",
-"S. c #B9AB8D",
-"D. c #BCAD93",
-"F. c #C0AF99",
-"G. c #C3B39C",
-"H. c #C9B99E",
-"J. c #CABAA2",
-"K. c #D4C2A9",
-"L. c #D9C8AC",
-"P. c #DFCCB3",
-"I. c #E3CEB5",
-"U. c #E6D3B8",
-"Y. c #EBD8BC",
-"T. c #80B4C7",
-"R. c #81B5C9",
-"E. c #82B8CC",
-"W. c #8AC6DD",
-"Q. c #EFDCC0",
-"!. c #F2DFC3",
-"~. c #FAE6C9",
-"^. c #FCE8CB",
-"/. c #F5E2C5",
-"(. c None",
-/* pixels */
-"(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.X . (.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.> * (.(.(.(.(.(.(.(.(.(.(.: B ~ @.%... .H V (.(.(.(.(.(.(.(.(.(.(.(.(.(.",
-"(.(.(.(.(.% 3.< (.(.+ (.(.(.(.(.(.(.V *.;.;.;.;.;.;.X.' [ q (.(.(.(.(.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.& (.(.(.3.& (.(.(.(.X ~ ;.;.;.=.Q v W O.;.&.{ ' C (.(.(.(.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.+ T c.3.+ (.(.(.~ ;.;.;.=.x R.V.R.b._ ;.&.' ' Z (.(.(.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.& (.(.* 3.* (.(.(.D ;.;.;.;.Q R.M D.Z.M.m ;.;.X.' { 0 (.(.(.(.(.(.(.(.(.",
-"(.(.(.(.(.+ 3.2 (.(.O (.(.(.: -.;.;.;.;.v B.J.~.^.g.M.| ;.;. .' ] (.(.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.= * (.(.(.(.(.(.L ;.;.;.;.;.^ E.Z.~.Q.v.b.#.;.;.X.' ' 0 (.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.. %.;.;.;.;.;.` M.M.g.v.E.x ;.;.;.;.' ' F (.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.r ;.;.;.;.;.;.;.! M M.b.x #.;.;.;.;. .' ] (.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.q ;.;.;.-.;.;.;.;.%.O.+.;.;.;.;.;.;...' { (.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.: B B B V S S S S S S S S S S S S S S Z F G C (.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.} ;.;.;.;.-.;.;.;.;.;.;.;.;.-.;.;.;.;.;.&.' ' C (.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.=.;.;.*.~ J J K J J J J J :.:.` J J K +.-.' ' F (.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.;.;.+.T j.l.l.l.l.l.l.l.l.c.c.z.l.l.k.r.R ' ' F (.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.-.-.I z.8.>.>.>.>.>.>.>.>.>.>.>.>.>.,.s.y.[ ' F (.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.@.-.U u.M.W.W.W.W.W.W.W.W.W.W.W.W.W.W.z z.P ' F (.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.q o.5.t.V.W.W.W.W.W.W.W.W.W.W.W.W.W.W.<.c.L [ o (.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.(.2.t.V.W.W.W.W.W.W.V.T.W.W.W.W.W.W.<.c.& (.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.(.3.p.V.W.W.W.W.W.1.r.0.m.W.W.W.W.W.<.c.= (.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.(.2 k.N N.N.N.N.v.w.c.c.,.m.N.N.N.n.,.l.& (.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.(.* h.k.h.h.s.h.j.z.,.0.c.h.i.i.h.h.z.< (.(.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.> < f 7.9.q.9.d.!.L.8.9.9.q.f 2 3 (.(.(.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.u Y.^.^.^.^.^.^.^.^.A.(.(.(.(.(.(.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.h Q.^.^.^.^.^.^.D.(.(.(.(.(.(.(.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.u P.^.^.^.~.A.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.  (.(.u ) A.f.h t (.@   (.(.(.(.(.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.@ ( t (.D.^.J.U.^.g   c c (.(.(.(.(.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.7 S 4 ( n ) !.^.^.^.^.L.u ( ) : w $ (.(.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.(.r =.;.t ( ( j I.^.^.^.^.A.c ( ( 8 -.K 7 (.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.7 *.;.*.t ( ( ( j f.K.G./ c ( ( ( 8 ;.;.S (.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.D ;.;.e c ( ( ( ( n j j / ( ( ( ( u o.;.=.7 (.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.~ ;.=.5 ( ( ( ) j c l l c h ( ( ( / 8 ;.;.e (.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.} ;.K v ( ( ( c /.K.S.D.U.P.j ( ( ( t ;.;.e (.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.} ;.> d d d a P.d.p.k.h.6.S.g.s d d < R ;.e (.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.} ;.d c.c.c.8.H.a.c.c.c.z.7.P.r.c.c.c.I ;.e (.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.} ;.i p p p g.d.z.z.c.c.c.s.G.h p p 6 O.;.e (.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.7 9 4 @ 4 @ 5 6 > - > > - - 6 4 4 4 # 7 7 $ (.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.",
-"(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(.(."
-};
