--- srcs/img/w.xpm
@@ -1,113 +1,0 @@
-/* XPM */
-static char *result[] = {
-/* columns rows colors chars-per-pixel */
-"64 64 43 1 ",
-"  c #778B9B",
-". c #778A9C",
-"X c #778B9C",
-"o c #788B9A",
-"O c #788A9B",
-"+ c #788B9B",
-"@ c #788A9C",
-"# c #788B9C",
-"$ c #7A8E9E",
-"% c #7B8E9F",
-"& c #7C8FA0",
-"* c #7D8FA1",
-"= c #7C90A0",
-"- c #7E91A2",
-"; c #7F92A2",
-": c #7F92A3",
-"> c #7F93A3",
-", c #8093A4",
-"< c #8193A5",
-"1 c #8194A5",
-"2 c #8396A7",
-"3 c #8497A7",
-"4 c #8598A9",
-"5 c #8698AA",
-"6 c #889BAC",
-"7 c #899CAC",
-"8 c #8A9CAD",
-"9 c #899CAE",
-"0 c #8B9EAE",
-"q c #8C9FAF",
-"w c #8C9FB0",
-"e c #8EA0B1",
-"r c #8EA0B2",
-"t c #92A4B5",
-"y c #95A7B9",
-"u c #99ABBD",
-"i c #9AACBD",
-"p c #9BACBE",
-"a c #9CAEC0",
-"s c #A0B1C3",
-"d c #A5B6C9",
-"f c #B0C1D4",
-"g c None",
-/* pixels */
-"gggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggg",
-"gggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggg",
-"gggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggg",
-"gggO          ######################++X#####################.ggg",
-"ggg $222222222&#22222222222222222222<#=22222222222222222222$#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg#&rrrrrrrrr,#,,,,,,,,,-%:,,,,,,,,:#%,,,,,,,,,,5rrrrrrrrr*#ggg",
-"ggg##########################################################ggg",
-"ggg#2ffffffffffffffffffffq#dffffffffffffffffffff#rfffffffff2 ggg",
-"ggg#2ffffffffffffffffffffq#dffffffffffffffffffff#rfffffffff2 ggg",
-"ggg#2ffffffffffffffffffffq#dffffffffffffffffffff#rfffffffff2 ggg",
-"ggg#2ffffffffffffffffffffq#dffffffffffffffffffff#rfffffffff2 ggg",
-"ggg#2ffffffffffffffffffffq#dffffffffffffffffffff#rfffffffff2 ggg",
-"ggg#2ffffffffffffffffffffq#dffffffffffffffffffff#rfffffffff2 ggg",
-"ggg#2ffffffffffffffffffffq#dffffffffffffffffffff#rfffffffff2 ggg",
-"ggg#2ffffffffffffffffffffq#dffffffffffffffffffff#rfffffffff2 ggg",
-"ggg#2ffffffffffffffffffffq#dffffffffffffffffffff#rfffffffff2 ggg",
-"ggg#:iiiiiiiiitqqqqqqqqqq>#6qqqqqqqq037qqqqqqqqq#4iiiiiiiii;@ggg",
-"ggg##########################################################ggg",
-"gggo1ddddddddd9#dddddddddddddddddddda#6dddddddddddddddddddd<#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"gggo1ddddddddd8#aaaaaaaaauypaaaaaaaay#4aaaaaaaaaasddddddddd<#ggg",
-"ggg##########################################################ggg",
-"ggg#:iiiiiiiiiiiiiiiiiiiiw#eiiiiiiiiiiiiiiiiiiii#4iiiiiiiii;@ggg",
-"ggg#2ffffffffffffffffffffi#affffffffffffffffffff#rfffffffff2 ggg",
-"ggg#2ffffffffffffffffffffi#affffffffffffffffffff#rfffffffff2 ggg",
-"ggg#2ffffffffffffffffffffi#affffffffffffffffffff#rfffffffff2 ggg",
-"ggg#2ffffffffffffffffffffi#affffffffffffffffffff#rfffffffff2 ggg",
-"ggg#2ffffffffffffffffffffi#affffffffffffffffffff#rfffffffff2 ggg",
-"ggg#2ffffffffffffffffffffi#affffffffffffffffffff#rfffffffff2 ggg",
-"ggg#2ffffffffffffffffffffi#affffffffffffffffffff#rfffffffff2 ggg",
-"ggg#2ffffffffffffffffffffi#affffffffffffffffffff#rfffffffff2 ggg",
-"ggg#2ffffffffffffffffffffi#affffffffffffffffffff#rfffffffff2 ggg",
-"ggg##########################################################ggg",
-"ggg#&rrrrrrrrr,#rrrrrrrrrrrrrrrrrrrr8#,rrrrrrrrrrrrrrrrrrrr*#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg 2fffffffffr#ffffffffffffffffffffd#qffffffffffffffffffff2#ggg",
-"ggg $222222222&#22222222222222222222<#=22222222222222222222$#ggg",
-"gggO          ######################++X#####################.ggg",
-"gggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggg",
-"gggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggg",
-"gggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggg"
-};
