+++ srcs/parse_map/init_map_info.c
@@ -6,7 +6,7 @@
 /*   By: jayoon <jayoon@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/17 13:04:54 by jayoon            #+#    #+#             */
-/*   Updated: 2022/11/10 21:43:40 by jayoon           ###   ########.fr       */
+/*   Updated: 2022/11/17 16:50:18 by jayoon           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
